Cleaned up Maven build file and restructured source files according to Maven documentation for WAR targets.
