+++ src/main/java/ch/hsr/bll/CDAR_SelfInformation.java
@@ -1,0 +1,24 @@
+package ch.hsr.bll;
+
+import java.util.LinkedHashMap;
+
+/**
+ * @author dzigerli@hsr.ch (Daniel Zigerlig)
+ */
+public abstract class CDAR_SelfInformation {
+    protected LinkedHashMap<String, String> information;
+
+    /**
+     * used for (re-)initialization of the class
+     */
+    protected abstract void init();
+
+    /**
+     * @returns name of information as string
+     */
+    public abstract String toString();
+
+    public LinkedHashMap<String, String> getInformation() {
+        return this.information;
+    }
+}
