+++ src/main/java/ch/hsr/dal/CDAR_SelfInformationProperties.java
@@ -4,8 +4,7 @@ import java.util.LinkedHashMap;
 
 /**
  * @author dzigerli@hsr.ch (Daniel Zigerlig)
- *
  */
 public interface CDAR_SelfInformationProperties {
-	public LinkedHashMap<String, String> getProperties();
+    public LinkedHashMap<String, String> getProperties();
 }
