--- war/WEB-INF/.gitignore
@@ -1,1 +1,0 @@
-/classes/
